Made package Python 3 compatible
Python 3 uses absolute import by default, so it cannot find the `clev` package in the global scope.
Adding `.` forces Python 3 to use relative import, and it should find `clev` in the current directory. This is also backwards compatible with Python 2.

I've tested this locally in both Python 2 and 3, but we should test after publishing to PyPI just in case.